+++ frontend/src/Api/Candidate/paths.ts
@@ -1,0 +1,8 @@
+
+
+export default {
+    me: '/candidates/me',
+    updateCandidate: '/candidates/me',
+    uploadAvatar: '/candidates/me/avatar',
+    uploadResume: '/candidates/me/resume',
+}
