+++ frontend/src/Api/Company/paths.ts
@@ -1,0 +1,7 @@
+
+
+export default {
+    me: '/companies/me',
+    updateCompany: '/companies/me',
+    uploadLogo: '/companies/me/logo',
+}
