--- frontend/src/Api/Profile/paths.ts
@@ -1,10 +1,0 @@
-
-
-export default {
-    candidateMe: '/candidates/me',
-    updateCandidate: '/candidates/me',
-    companyMe: '/companies/me',
-    updateCompany: '/companies/me',
-    uploadCandidateAvatar: '/candidates/me/avatar',
-    uploadCompanyLogo: '/companies/me', // TODO: обновить когда будет готов бек
-}
