+++ frontend/src/Api/index.ts
@@ -1,4 +1,5 @@
 export {default as Auth} from './Auth/Auth'
 export {default as User} from './User/User'
-export {default as Profile} from './Profile/Profile'
+export {default as Candidate} from './Candidate/Candidate'
+export {default as Company} from './Company/Company'
 export {default as Vacancies} from './Vacancies/Vacancies'
