+++ frontend/src/pages/profile/Hooks/useCandidateProfile.ts
@@ -4,6 +4,6 @@ import * as Api from '@/Api';
 export const useCandidateProfile = () => {
     return useQuery({
         queryKey: ['candidate-me'],
-        queryFn: () => Api.Profile.candidateMe(),
+        queryFn: () => Api.Candidate.me(),
     })
 }
