+++ frontend/src/pages/profile/Hooks/useCompanyProfile.ts
@@ -4,6 +4,6 @@ import * as Api from '@/Api';
 export const useCandidateProfile = () => {
     return useQuery({
         queryKey: ['company-me'],
-        queryFn: () => Api.Profile.companyMe(),
+        queryFn: () => Api.Company.me(),
     })
 }
